+++ src/main/java/next/web/user/ListUserServlet.java
@@ -1,4 +1,4 @@
-package next.web;
+package next.web.user;
 
 import java.io.IOException;
